core: top-level-sheet-pane: make mirrored-sheet-mixin a superclass
A top-level pane almost always requires a mirror. Instead of making all frame
managers separately add a mixin we subclass MIRRORED-SHEET-MIXIN directly.

As of exceptions - an exception would be when we create a graft that is in
fact a window - we could then reuse the graft's mirror. That said this change
is not an issue in this scenario, because:

- since it is a pane then the frame manager that does such sheningans may
return a different (i.e unmirrored) pane from find-pane-for-frame

- realize-mirror may blatantly assign graft's mirror to the sheet - that would
probably reveal some buggy assumptions that the sheet native region is the
same as the mirror for mirrored sheets, but oh well, let's leave it to /them/

- top-level-sheet-mixin is not mirrored, so it is possible to construct top
level panes that are not mirrored, just call them differently (!) - that would
probably reveal even more buggy assumptions that the top-level-sheet-pane is
always the frame's top level sheet, ditto, leave it to future developers