extensions: render: remove MAYBE-{TRANSFORMATION,REGION}
These functions were introduced to hide some architectural problem with the
module. These problems are gone now, so we directly call appropriate
functions.

Also fix a bug where a native transformation was used for drawing - while this
issue is not visible most of the time (due to recording), all drawing should
use the device transformation.