extensions: render: remove unused function for drawing text
The function AA-CELLS-ALPHA-SWEEP/RECTANGLE was never called in the codebase.
In the past it was meant for text rendering (but apparently was never used).

Bring back AA-FILL-ALPHA-PATHS that was removed long time ago (because it was
not called), that called AA-CELLS-ALPHA-SWEEP/RECTANGLE and make it call the
"usual" trampoline AA-CELLS-SWEEP/RECTANGLE and add appropriate branch to the
latter.

This new code path is commented out with the function AA-RENDER-ALPHA-DRAW-FN,
to make it clear that they are not used. I'm leaving them for now because
maybe they'll prove useful when we'll try to refactor text drawing (which is
awfully slow).