core: space-requirements: remove "helper" macros
Remove macros sr{,min,max}-{width,height}. They were used only in a limited
number of places and always re-accessed the requirement from the pane. This
did not make the program readable. During removal a typo was found in
ALLOCATE-SPACE specialized on the GRID-PANE, namely SR-MIN-WIDTH was used for
MAX-WIDTH where SR-MAX-WIDTH should be used.